+++ petnetizen_feeder/__init__.py
@@ -43,7 +43,7 @@ from .feeder import (
     discover_feeders,
 )
 
-__version__ = "0.5.5"
+__version__ = "0.5.6"
 __all__ = [
     "ConnectionFactory",
     "FeederDevice",